Replace codeparser WIP
